Extract CardLabel into card_label.dart
