+++ lib/src/sdk/cards/cards.dart
@@ -3,5 +3,6 @@ export 'attachment_id.dart';
 export 'card.dart';
 export 'card_client.dart';
 export 'card_id.dart';
+export 'card_label.dart';
 export 'card_models.dart';
 export 'trello_card.dart';
